Update Instant_Articles_Post->_get_the_content() to validate $post->ID before using it
Fixes the following error which blocks saving contact forms in Contact Form 7:

`Notice: Trying to get property of a non-object in /.../.../fb-instant-articles/class-instant-articles-post.php on line 299`

Whatever this code is trying to do, I don't think it intends to throw an error when `$post->ID` is empty. 

This edit replicates the old behavior but without throwing a notice:  If `$post` is undefined OR the ID doesn't match `$this->_post` then we run setup_postdata to ensure globals reflect `$this->post_id`. 

The undefined `$post` state was previously implied rather than specified, but I can't see how the old code would behave differently from this edit.