Added tranformation warning indicator column
This branch adds a column to wp_edit pages called FBIA Status. The column shows a green light if the post was transformed into a Facebook Instant article with no problems, a yellow light if there were warnings, and a red light if the post was empty after trasnformation.

There is a checkbox in the FBIA Options page that defaults to false, so that editors can decide whether to show the column or not. The checkbox is placed alongside the "Publish with warnings" option for good grouping of concerns.

A new stylesheet has been added to contain the styling rules for the warning indicators. Since the indicators are very minimal, just red, green, and yellow lights, an informative title text is available on mouseover to help with accessibility.