Update Readme re:non-minimal readership
Now that it is impossible to use this plugin with a site that has ["minimal readership" on Facebook](https://github.com/Automattic/facebook-instant-articles-wp/issues/925#issuecomment-385243511) it is vital that the plugin warns developers not to set it up unless their site already meets the guidelines. This PR is a first step to warn people that after all their hard work getting their site to validate as IA they will still likely be blocked because of minimal readership.