Reduce plugin's cache burden, remote requests
Cached or not, the result of the remote request is only used to determine if
`instant_articles_embed_get_html()` should be called, which itself returns
early when called outside of the transformation process. If a post is never
published to Instant Articles, neither the remote request nor its cached result
are necessary; due to the filter this is hooked to, it will be called regardless.